requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
langchain-text-splitters
faiss-cpu
pymupdf
sentence-transformers
fastapi
uvicorn
python-dotenv
python-multipart
